chore: new interface, fix build (#18)
<!-- Remove all rows that don't apply to this PR --->
| Type | Description | Icon | Changelog |
| ---------- |
---------------------------------------------------------------------------|
:--: | :-------: |
| `chore` | Internal stuff (feat/fix/debt/etc...) | 👷 | |

-------------------------------------------------------------------------------------------------------------

### 🛠️ Changes:
<!-- Describe your changes in a brief, bullet list --->
- fix build output (new versions changed the extensions on the output
files)
- use PascalCase for enums and enum keys
- use vue's script setup to define the components
- add a simplified version of the SplitPanel that doesn't require as
much boilerplate
- add the ability to auto constrain the panel to the content scroll
dimensions

### 📢 Additional Info:
<!-- Give some more context... Why is this needed? --->
This should make splitpanel a bit easier to use. Most of the
configuration should be done automatically if you use the
`SplitPanelView`/`SplitPanelItem`. This also adds the ability to auto
constrain the panel size based on the scroll height of the panel, and
adds the ability to enable/disable panels. Those two features are useful
if you want to display something with a fixed size and you don't want
that panel to be any size in between. It's either on or off.

### 📚 Bookkeeping:

**Testing (if applicable)**:
- [ ] Ran/wrote unit tests for this

**Checklist**
- [x] Assigned PR to myself
- [x] Added at least 1 person on the team as reviewer
- [ ] Release Notes: PRs types that have the :spiral_notepad: next to
them also require release notes to be added to the `CHANGELOG.md`